+++ models/NewsCategoryModel.php
@@ -28,6 +28,32 @@ class NewsCategoryModel extends \Model
     protected static $strTable = 'tl_news_category';
 
     /**
+     * Get the target page
+     *
+     * @return \PageModel|null
+     */
+    public function getTargetPage()
+    {
+        $pageId = $this->jumpTo;
+
+        // Inherit the page from parent if there is none set
+        if (!$pageId) {
+            $pid = $this->pid;
+
+            do {
+                $parent = static::findByPk($pid);
+
+                if ($parent !== null) {
+                    $pid = $parent->pid;
+                    $pageId = $parent->jumpTo;
+                }
+            } while ($pid && !$pageId);
+        }
+
+        return \PageModel::findByPk($pageId);
+    }
+
+    /**
      * Find published news categories by their archives
      *
      * @param array $arrArchives An array of archives